fix(ui): consistent use of formatValue for colorbar and hover readout
